Add files_from parameter to backup command (#190)
## Contributing process

- [x] I understand that this is [some guy's personal hobby
project](https://github.com/mtlynch/resticpy#resticpys-scope-and-future)
and reviews are on a best-effort basis
- [x] I've read the [contribution
guidelines](../blob/master/CONTRIBUTING.md)

## What type of PR is this? (check all applicable)

- [x] Refactor
- [x] Feature
- [ ] Bugfix
- [ ] Optimization
- [ ] Documentation Update

## Description

This PR adds an `files_from` parameter to the backup command.
`files_from` takes, as `exclude_files`, a list of files as an argument.
This allows to specify one (or multiple) file that contains a file list.

It will call restic with the `--files-from` option. 

In addition the following changes were necessary: 
- Changing `paths` to an optional argument with default None. This
allows to call backup with include_files, only.
- Ensure that either `paths` or `files_from_files` (or both) are
specified and not None. Raise `ValueError` in case both are None.
- Refactoring command extension with list into helper function.
circlci:build was failing with too many if clauses in function.

## Did you update the [API documentation](../blob/master/docs)?

- [x] Yes
- [ ] This change does not require documentation changes
- [ ] I need help updating documentation

## Have you added or updated tests?

- [x] Yes, I added unit tests
- [ ] Yes, I added [end-to-end tests](../blob/master/e2e/)
- [ ] No, and this is why: _please replace this line with details on why
tests have not been included_
- [ ] I need help writing tests